Relocate local spmd_block in lcos::local namespace
