+++ hpx/parallel/spmd_block.hpp
@@ -7,44 +7,16 @@
 #define HPX_PARALLEL_SPMD_BLOCK_HPP
 
 #include <hpx/lcos/future.hpp>
-#include <hpx/lcos/local/barrier.hpp>
+#include <hpx/lcos/local/spmd_block.hpp>
 #include <hpx/parallel/execution_policy.hpp>
 #include <hpx/traits/is_execution_policy.hpp>
 
-#include <boost/range/irange.hpp>
-
 #include <cstddef>
-#include <functional>
-#include <memory>
 #include <type_traits>
-#include <utility>
 #include <vector>
 
 namespace hpx { namespace parallel { inline namespace v2
 {
-    namespace detail
-    {
-        template <typename T>
-        struct extract_first_parameter
-        {};
-
-        // Specialization for lambdas
-        template <typename ClassType, typename ReturnType>
-        struct extract_first_parameter<ReturnType(ClassType::*)() const>
-        {
-            using type = std::false_type;
-        };
-
-        // Specialization for lambdas
-        template <typename ClassType,
-            typename ReturnType, typename Arg0, typename... Args>
-        struct extract_first_parameter<
-            ReturnType(ClassType::*)(Arg0, Args...) const>
-        {
-            using type = typename std::decay<Arg0>::type;
-        };
-    }
-
     /// The class spmd_block defines an interface for launching
     /// multiple images while giving handles to each image to interact with
     /// the remaining images. The \a define_spmd_block function templates create
@@ -52,101 +24,20 @@ namespace hpx { namespace parallel { inline namespace v2
     /// separate thread. A temporary spmd block object is created and diffused
     /// to each image. The constraint for the lambda given to the
     /// define_spmd_block function is to accept a spmd_block as first parameter.
-    struct spmd_block
-    {
-        explicit spmd_block(std::size_t num_images, std::size_t image_id,
-            hpx::lcos::local::barrier & barrier)
-        : num_images_(num_images), image_id_(image_id), barrier_(barrier)
-        {}
-
-        // Note: spmd_block class is movable/move-assignable
-        // but not copyable/copy-assignable
-
-        spmd_block(spmd_block &&) = default;
-        spmd_block(spmd_block const &) = delete;
-
-        spmd_block & operator=(spmd_block &&) = default;
-        spmd_block & operator=(spmd_block const &) = delete;
-
-        std::size_t get_num_images() const
-        {
-            return num_images_;
-        }
-
-        std::size_t this_image() const
-        {
-            return image_id_;
-        }
-
-        void sync_all() const
-        {
-           barrier_.get().wait();
-        }
-
-    private:
-        std::size_t num_images_;
-        std::size_t image_id_;
-        mutable std::reference_wrapper<hpx::lcos::local::barrier> barrier_;
-    };
-
-    namespace detail
-    {
-        template <typename F>
-        struct spmd_block_helper
-        {
-            mutable std::shared_ptr<hpx::lcos::local::barrier> barrier_;
-            typename std::decay<F>::type f_;
-            std::size_t num_images_;
-
-            template <typename ... Ts>
-            void operator()(std::size_t image_id, Ts && ... ts) const
-            {
-                spmd_block block(num_images_, image_id, *barrier_);
-                hpx::util::invoke(
-                    f_, std::move(block), std::forward<Ts>(ts)...);
-            }
-        };
-    }
+    using spmd_block = hpx::lcos::local::spmd_block;
 
     // Asynchronous version
     template <typename ExPolicy, typename F, typename ... Args,
         typename = typename std::enable_if<
             hpx::parallel::v1::is_async_execution_policy<ExPolicy>::value>::type
         >
-    std::vector<hpx::future<void>>
+    inline std::vector<hpx::future<void>>
     define_spmd_block(ExPolicy && policy,
         std::size_t num_images, F && f, Args && ... args)
     {
-        static_assert(
-            parallel::execution::is_execution_policy<ExPolicy>::value,
-            "parallel::execution::is_execution_policy<ExPolicy>::value");
-
-        using ftype = typename std::decay<F>::type;
-
-        using first_type =
-            typename hpx::parallel::v2::detail::extract_first_parameter<
-                        decltype(&ftype::operator())>::type;
-
-        using executor_type =
-            typename hpx::util::decay<ExPolicy>::type::executor_type;
-
-        static_assert(std::is_same<spmd_block, first_type>::value,
-            "define_spmd_block() needs a lambda that " \
-            "has at least a spmd_block as 1st argument");
-
-        std::shared_ptr<hpx::lcos::local::barrier> barrier
-            = std::make_shared<hpx::lcos::local::barrier>(num_images);
-
-        return
-            hpx::parallel::executor_traits<
-                    typename std::decay<executor_type>::type
-                >::bulk_async_execute(
-                    policy.executor(),
-                    detail::spmd_block_helper<F>{
-                        barrier, std::forward<F>(f), num_images
-                    },
-                    boost::irange(std::size_t(0), num_images),
-                    std::forward<Args>(args)...);
+        return hpx::lcos::local::define_spmd_block(
+            std::forward<ExPolicy>(policy), num_images,
+                std::forward<F>(f), std::forward<Args>(args)...);
     }
 
     // Synchronous version
@@ -154,45 +45,20 @@ namespace hpx { namespace parallel { inline namespace v2
         typename = typename std::enable_if<
             !hpx::parallel::v1::is_async_execution_policy<ExPolicy>::value>::type
         >
-    void
+    inline void
     define_spmd_block(ExPolicy && policy,
         std::size_t num_images, F && f, Args && ... args)
     {
-        static_assert(
-            parallel::execution::is_execution_policy<ExPolicy>::value,
-            "parallel::execution::is_execution_policy<ExPolicy>::value");
-
-        using ftype = typename std::decay<F>::type;
-
-        using first_type =
-            typename hpx::parallel::v2::detail::extract_first_parameter<
-                        decltype(&ftype::operator())>::type;
-
-        using executor_type =
-            typename hpx::util::decay<ExPolicy>::type::executor_type;
-
-        static_assert(std::is_same<spmd_block, first_type>::value,
-            "define_spmd_block() needs a lambda that " \
-            "has at least a spmd_block as 1st argument");
-
-        std::shared_ptr<hpx::lcos::local::barrier> barrier
-            = std::make_shared<hpx::lcos::local::barrier>(num_images);
-
-        hpx::parallel::executor_traits<
-                typename std::decay<executor_type>::type
-            >::bulk_execute(
-                policy.executor(),
-                detail::spmd_block_helper<F>{
-                    barrier, std::forward<F>(f), num_images
-                },
-                boost::irange(std::size_t(0), num_images),
-                std::forward<Args>(args)...);
+        hpx::lcos::local::define_spmd_block(
+            std::forward<ExPolicy>(policy), num_images,
+                std::forward<F>(f), std::forward<Args>(args)...);
     }
 
     template <typename F, typename ... Args>
-    void define_spmd_block(std::size_t num_images, F && f, Args && ... args)
+    inline void define_spmd_block(std::size_t num_images,
+        F && f, Args && ... args)
     {
-        define_spmd_block(parallel::execution::par,
+        hpx::lcos::local::define_spmd_block(parallel::execution::par,
             num_images, std::forward<F>(f), std::forward<Args>(args)...);
     }
 }}}
